pyproject.toml
[tool.poetry]
name = "dog-shelter"
version = "0.0.1"
description = "Project for showcase FastAPI functionality"
authors = ["dqmis <dom.seputis@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.*"
fastapi = "^0.85.1"
uvicorn = "^0.19.0"
mysql-connector-python = "^8.0.31"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^22.6.0"
flake8 = "^4.0.1"
pre-commit = "^2.20.0"
pytest = "^7.1.2"
safety = "^1.10.3"
mypy = "^0.971"
bandit = "^1.7.4"
isort = "^5.10.1"
jupyterlab = "^3.4.7"

[tool.flake8]
exclude = '''
/(
    \.git
  | \.hg
  | \.eggs
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.vscode
  | _build
  | buck-out
  | build
  | dist
  | condaenv
)/
'''
max_line_length = 99
max-complexity = 18

[tool.black]
line-length = 99
preview = true
include = '/project$|\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.eggs
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.vscode
  | _build
  | buck-out
  | build
  | dist
  | condaenv
)/
'''

[tool.isort]
profile = "black"
line_length = 99
